fix zmq reception of stream structure
TODO: fix issue with it failing to deserialize

moved most all the stream struct's and other unneeded code into
src/lib.rs.
