+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod stream_data;
 
+use std::sync::Arc;
 use std::time::{SystemTime, UNIX_EPOCH};
 
 // Function to get the current Unix timestamp in milliseconds
@@ -8,4 +9,21 @@ pub fn current_unix_timestamp_ms() -> Result<u64, &'static str> {
         .duration_since(UNIX_EPOCH)
         .map(|d| d.as_millis() as u64)
         .map_err(|_| "System time is before the UNIX epoch")
+}
+
+// Print a hexdump of the packet
+pub fn hexdump(packet_arc: &Arc<Vec<u8>>, packet_offset: usize, packet_len: usize) {
+    let packet = &packet_arc[packet_offset..packet_offset + packet_len];
+    println!("--------------------------------------------------");
+    // print in rows of 16 bytes
+    let mut packet_dump = String::new();
+    for (i, chunk) in packet.iter().take(packet_len).enumerate() {
+        if i % 16 == 0 {
+            packet_dump.push_str(&format!("\n{:04x}: ", i));
+        }
+        packet_dump.push_str(&format!("{:02x} ", chunk));
+    }
+    println!("{}", packet_dump);
+    println!("");
+    println!("--------------------------------------------------");
 }
