+++ src/stream_data.rs
@@ -5,7 +5,7 @@
 */
 
 use crate::current_unix_timestamp_ms;
-use log::error;
+use log::{debug, error};
 use serde::{Deserialize, Serialize};
 use std::{fmt, sync::Arc};
 
@@ -347,4 +347,161 @@ impl Tr101290Errors {
             cat_errors: 0,
         }
     }
+}
+
+// TR 101 290 Priority 1 Check
+pub fn tr101290_p1_check(packet: &[u8], errors: &mut Tr101290Errors) {
+    // p1
+    if packet[0] != 0x47 {
+        errors.sync_byte_errors += 1;
+    }
+
+    // TODO: ... other checks, updating the respective counters ...
+}
+
+// TR 101 290 Priority 2 Check
+pub fn tr101290_p2_check(packet: &[u8], errors: &mut Tr101290Errors) {
+    // p2
+
+    if (packet[1] & 0x80) != 0 {
+        errors.transport_error_indicator_errors += 1;
+    }
+    // TODO: ... other checks, updating the respective counters ...
+}
+
+// Implement a function to extract PID from a packet
+pub fn extract_pid(packet: &[u8]) -> u16 {
+    if packet.len() < TS_PACKET_SIZE {
+        return 0; // Packet size is incorrect
+    }
+
+    let transport_error = (packet[1] & 0x80) != 0;
+    if transport_error {
+        return 0xFFFF; // Packet has a transport error
+    }
+
+    // Extract PID from packet
+    ((packet[1] as u16 & 0x1F) << 8) | packet[2] as u16
+}
+
+// Story the PAT packet and PMT PID
+pub struct PmtInfo {
+    pub pid: u16,
+    pub packet: Vec<u8>,
+}
+
+// Helper function to parse PAT and update global PAT packet storage
+pub fn parse_and_store_pat(packet: &[u8]) -> PmtInfo {
+    let pat_entries = parse_pat(packet);
+    let mut pmt_info = PmtInfo {
+        pid: 0xFFFF,
+        packet: Vec::new(),
+    };
+    pmt_info.packet = packet.to_vec();
+
+    // Assuming there's only one program for simplicity, update PMT PID
+    if let Some(first_entry) = pat_entries.first() {
+        pmt_info.pid = first_entry.pmt_pid;
+    }
+    pmt_info
+}
+
+pub fn parse_pat(packet: &[u8]) -> Vec<PatEntry> {
+    let mut entries = Vec::new();
+
+    // Check for minimum packet size
+    if packet.len() < TS_PACKET_SIZE {
+        return entries;
+    }
+
+    // Check if Payload Unit Start Indicator (PUSI) is set
+    let pusi = (packet[1] & 0x40) != 0;
+    if !pusi {
+        // If Payload Unit Start Indicator is not set, this packet does not start a new PAT
+        return entries;
+    }
+
+    let adaptation_field_control = (packet[3] & 0x30) >> 4;
+    let mut offset = 4; // start after TS header
+
+    // Check for adaptation field and skip it
+    if adaptation_field_control == 0x02 || adaptation_field_control == 0x03 {
+        let adaptation_field_length = packet[4] as usize;
+        offset += 1 + adaptation_field_length; // +1 for the length byte itself
+    }
+
+    // Pointer field indicates the start of the PAT section
+    let pointer_field = packet[offset] as usize;
+    offset += 1 + pointer_field; // Skip pointer field
+
+    // Now, 'offset' points to the start of the PAT section
+    while offset + 4 <= packet.len() {
+        let program_number = ((packet[offset] as u16) << 8) | (packet[offset + 1] as u16);
+        let pmt_pid = (((packet[offset + 2] as u16) & 0x1F) << 8) | (packet[offset + 3] as u16);
+
+        // Only add valid entries (non-zero program_number and pmt_pid)
+        if program_number != 0 && pmt_pid != 0 && pmt_pid < 0x1FFF && program_number < 100 {
+            entries.push(PatEntry {
+                program_number,
+                pmt_pid,
+            });
+        }
+
+        debug!(
+            "ParsePAT: Program Number: {} PMT PID: {}",
+            program_number, pmt_pid
+        );
+
+        offset += 4; // Move to the next PAT entry
+    }
+
+    entries
+}
+
+pub fn parse_pmt(packet: &[u8]) -> Pmt {
+    let mut entries = Vec::new();
+    let program_number = ((packet[8] as u16) << 8) | (packet[9] as u16);
+
+    // Calculate the starting position for stream entries
+    let section_length = (((packet[6] as usize) & 0x0F) << 8) | packet[7] as usize;
+    let program_info_length = (((packet[15] as usize) & 0x0F) << 8) | packet[16] as usize;
+    let mut i = 17 + program_info_length; // Starting index of the first stream in the PMT
+
+    debug!(
+        "ParsePMT: Program Number: {} PMT PID: {} starting at position {}",
+        program_number,
+        extract_pid(packet),
+        i
+    );
+    while i + 5 <= packet.len() && i < 17 + section_length - 4 {
+        let stream_type = packet[i];
+        let stream_pid = (((packet[i + 1] as u16) & 0x1F) << 8) | (packet[i + 2] as u16);
+        let es_info_length = (((packet[i + 3] as usize) & 0x0F) << 8) | packet[i + 4] as usize;
+        i += 5 + es_info_length; // Update index to point to next stream's info
+
+        entries.push(PmtEntry {
+            stream_pid,
+            stream_type,
+        });
+        debug!(
+            "ParsePMT: Stream PID: {}, Stream Type: {}",
+            stream_pid, stream_type
+        );
+    }
+
+    Pmt { entries }
+}
+
+// Helper function to identify the video PID from the stored PAT packet and return the PID and codec
+pub fn identify_video_pid(pmt_packet: &[u8]) -> Option<(u16, Codec)> {
+    let pmt = parse_pmt(pmt_packet);
+    pmt.entries.iter().find_map(|entry| {
+        let codec = match entry.stream_type {
+            0x01..=0x02 => Some(Codec::MPEG2), // MPEG-2 Video
+            0x1B => Some(Codec::H264),         // H.264 Video
+            0x24 => Some(Codec::H265),         // H.265 Video
+            _ => None,
+        };
+        codec.map(|c| (entry.stream_pid, c))
+    })
 }
